requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pillow>=10.1.0
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
numpy>=1.26.0
scikit-learn>=1.3.0
requests>=2.31.0
aiofiles>=23.2.1 